requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.7.0
trl>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Data Processing
convokit>=3.0.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0

# Logging & Monitoring
wandb>=0.15.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

